Implement setting page and clean nav bar
